+++ src/connect4/search/default_encoding.c
@@ -57,20 +57,20 @@ COL, ROW, PLAYER, BOARD, var, level
 void initialise_variables(nodeindex_t (**X)[2][2], uint32_t width, uint32_t height) {
     // Second, cells: order column, row, player, board
     if (!ALLOW_ROW_ORDER || width >= height) {
-        for (int col = 0; col < width; col++) {
-            for (int row = 0; row < height; row++) {
-                for (int player = 0; player < 2; player++) {
-                    for (int board = 0; board < 2; board++) {
+        for (uint8_t col = 0; col < width; col++) {
+            for (uint8_t row = 0; row < height; row++) {
+                for (uint8_t player = 0; player < 2; player++) {
+                    for (uint8_t board = 0; board < 2; board++) {
                         X[col][row][player][board] = create_variable();
                     }
                 }
             }
         }
     } else {
-        for (int row = 0; row < height; row++) {
-            for (int col = 0; col < width; col++) {
-                for (int player = 0; player < 2; player++) {
-                    for (int board = 0; board < 2; board++) {
+        for (uint8_t row = 0; row < height; row++) {
+            for (uint8_t col = 0; col < width; col++) {
+                for (uint8_t player = 0; player < 2; player++) {
+                    for (uint8_t board = 0; board < 2; board++) {
                         X[col][row][player][board] = create_variable();
                     }
                 }
@@ -80,9 +80,9 @@ void initialise_variables(nodeindex_t (**X)[2][2], uint32_t width, uint32_t heig
 }
 
 void initialise_variable_sets(nodeindex_t (**X)[2][2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, variable_set_t* vars_board0, variable_set_t* vars_board1) {
-    for (int col = 0; col < width; col++) {
-        for (int row = 0; row < height; row++) {
-            for (int player = 0; player < 2; player++) {
+    for (uint8_t col = 0; col < width; col++) {
+        for (uint8_t row = 0; row < height; row++) {
+            for (uint8_t player = 0; player < 2; player++) {
                 add_variable(vars_board0, X[col][row][player][0]);
                 add_variable(vars_board1, X[col][row][player][1]);
             }
@@ -95,9 +95,9 @@ void initialise_variable_sets(nodeindex_t (**X)[2][2], nodeindex_t stm0, nodeind
 // computes transition relation for single move
 // a move is placing a red/yellow stone (player) ate cell (col, row)
 // this is BDD with 2*(2*width*height + 1) variables
-nodeindex_t trans_move(int col, int row, int player, int current_board, nodeindex_t stm0, nodeindex_t stm1, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height) {
-    int opponent = player == 0 ? 1 : 0;
-    int next_board = current_board == 0 ? 1 : 0;
+nodeindex_t trans_move(uint8_t col, uint8_t row, uint8_t player, uint8_t current_board, nodeindex_t stm0, nodeindex_t stm1, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height) {
+    uint8_t opponent = player == 0 ? 1 : 0;
+    uint8_t next_board = current_board == 0 ? 1 : 0;
     
     // precondition (depends variables on of current board)
     nodeindex_t pre;
@@ -130,12 +130,12 @@ nodeindex_t trans_move(int col, int row, int player, int current_board, nodeinde
 
     // frame: every other cell stays the same
     nodeindex_t frame = ONEINDEX;
-    for (int c = 0; c < width; c++) {
-        for (int r = 0; r < height; r++) {
+    for (uint8_t c = 0; c < width; c++) {
+        for (uint8_t r = 0; r < height; r++) {
             if ((c == col) && (r == row)) {
                 continue;
             }
-            for (int p = 0; p < 2; p++) {
+            for (uint8_t p = 0; p < 2; p++) {
                 // cell is occupied with red/yellow stone in next board
                 // if it is occupied with red/yellow stone in current board
                 frame = and(frame, iff(X[c][r][p][0], X[c][r][p][1]));
@@ -148,11 +148,11 @@ nodeindex_t trans_move(int col, int row, int player, int current_board, nodeinde
 }
 
 
-nodeindex_t get_trans(nodeindex_t (**X)[2][2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, int board) {
+nodeindex_t get_trans(nodeindex_t (**X)[2][2], nodeindex_t stm0, nodeindex_t stm1, uint32_t width, uint32_t height, uint8_t board) {
     nodeindex_t trans = ZEROINDEX;
-    for (int col = 0; col < width; col++) {
-        for (int row = 0; row < height; row++) {
-            for (int player = 0; player < 2; player++) {
+    for (uint8_t col = 0; col < width; col++) {
+        for (uint8_t row = 0; row < height; row++) {
+            for (uint8_t player = 0; player < 2; player++) {
                 trans = or(trans, trans_move(col, row, player, board, stm0, stm1, X, width, height));
             }
         }
@@ -165,9 +165,9 @@ nodeindex_t get_trans(nodeindex_t (**X)[2][2], nodeindex_t stm0, nodeindex_t stm
 nodeindex_t connect4_start(nodeindex_t stm0, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height) {
     nodeindex_t s = ONEINDEX;
     s = and(s, stm0);
-    for (int c = 0; c < width; c++) {
-        for (int r = 0; r < height; r++) {
-            for (int p = 0; p < 2; p++) {
+    for (uint8_t c = 0; c < width; c++) {
+        for (uint8_t r = 0; r < height; r++) {
+            for (uint8_t p = 0; p < 2; p++) {
                 s = and(s, not(X[c][r][p][0]));
             }
         }
@@ -177,7 +177,7 @@ nodeindex_t connect4_start(nodeindex_t stm0, nodeindex_t (**X)[2][2], uint32_t w
 
 // Subtracts all positions from current which are terminal, i.e. four in a row, column or diagonal
 // Subtraction is performed iteratively and also performs GC.
-nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level, bool subtract) {
+nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level, bool subtract) {
     nodeindex_t a;
     nodeindex_t intersection = ZEROINDEX;
 
@@ -185,11 +185,11 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     // COLUMN
     if (height >= 4) {
-        for (int col = 0; col < width; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 0; col < width; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a column for player starting at (col,row)
                 a = ONEINDEX;
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     a = and(a, X[col][row + i][player][board]);
                 }
                 if (subtract) {
@@ -218,11 +218,11 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     // ROW
     if (width >= 4) {
-        for (int row = 0; row < height; row++) {
-            for (int col = 0; col <= width - 4; col++) {
+        for (uint8_t row = 0; row < height; row++) {
+            for (uint8_t col = 0; col <= width - 4; col++) {
                 // encodes that there are four stones in a row for player starting at (col,row)
                 a = ONEINDEX;
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     a = and(a, X[col + i][row][player][board]);
                 }
                 if (subtract) {
@@ -244,11 +244,11 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
 
     if (height >= 4 && width >= 4) {
         // DIAG ascending
-        for (int col = 0; col <= width - 4; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 0; col <= width - 4; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a ascending diagonal for player starting at (col,row)
                 a = ONEINDEX;
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     a = and(a, X[col + i][row + i][player][board]);
                 }
                 if (subtract) {
@@ -268,11 +268,11 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
         }
 
         // DIAG descending
-        for (int col = 3; col < width; col++) {
-            for (int row = 0; row <= height - 4; row++) {
+        for (uint8_t col = 3; col < width; col++) {
+            for (uint8_t row = 0; row <= height - 4; row++) {
                 // encodes that there are four stones in a descending diagonal for player starting at (col+3,row)
                 a = ONEINDEX;
-                for (int i = 0; i < 4; i++) {
+                for (uint8_t i = 0; i < 4; i++) {
                     a = and(a, X[col - i][row + i][player][board]);
                 }
                 if (subtract) {
@@ -301,9 +301,9 @@ nodeindex_t connect4_subtract_or_intersect_term(nodeindex_t current, int board, 
     }
 }
 
-inline nodeindex_t connect4_subtract_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level) {
+inline nodeindex_t connect4_subtract_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level) {
     return connect4_subtract_or_intersect_term(current, board, player, X, width, height, gc_level, true);
 }
-inline nodeindex_t connect4_intersect_term(nodeindex_t current, int board, int player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level) {
+inline nodeindex_t connect4_intersect_term(nodeindex_t current, uint8_t board, uint8_t player, nodeindex_t (**X)[2][2], uint32_t width, uint32_t height, int gc_level) {
     return connect4_subtract_or_intersect_term(current, board, player, X, width, height, gc_level, false);
 }
